+++ game/230/MM_HighScore.c
@@ -333,7 +333,7 @@ LAB_OVR_230__800b3c78:
 		iVar7 = (8 - D230.highScore_transitionFrames[1]) * D230.highScore_horizontalMove[0] * 0x40;
 	}
 
-	u_long *ot = sdata->gGT->backBuffer->otMem.uiOT;
+	uint32_t *ot = sdata->gGT->backBuffer->otMem.uiOT;
 
 	if (((iVar7 != -0x200) && (iVar7 != 0x200)) && ((iVar4 != -0xd8 && (iVar4 != 0xd8))))
 	{
