+++ game/Display.c
@@ -151,7 +151,7 @@ void DISPLAY_Blur_Main(struct PushBuffer *pb, int strength)
 	u32 *prim = backBuffer->primMem.cursor;
 	u32 *nextPrim;
 	s8 cameraID;
-	u_long *ot;
+	uint32_t *ot;
 
 	cameraID = *(s8 *)&pb->cameraID;
 
@@ -200,7 +200,7 @@ void DISPLAY_Blur_Main(struct PushBuffer *pb, int strength)
 
 		ot = gGT->otSwapchainDB[gGT->swapchainIndex];
 		oldTag = *ot;
-		*ot = (u_long)CtrGpu_PrimToOTLink24(prim);
+		*ot = (uint32_t)CtrGpu_PrimToOTLink24(prim);
 
 		tile->dstX = pb->rect.x;
 		tile->dstY = pb->rect.y;
