+++ game/DrawConfetti/DrawConfetti.c
@@ -136,7 +136,7 @@ static int DrawConfetti_BoundsVisible(u32 sxy0, u32 sxy1, u32 sxy2, u32 sxy3, u3
 	return (s32)(bounds << 16) >= 0;
 }
 
-static void DrawConfetti_LinkPrimitive(POLY_F4 *poly, u_long *ot)
+static void DrawConfetti_LinkPrimitive(POLY_F4 *poly, uint32_t *ot)
 {
 	CtrGpu_LinkPacket24(ot, &poly->tag, poly, 0x05000000);
 }
@@ -148,7 +148,7 @@ void DrawConfetti(struct PushBuffer *pb, struct PrimMem *primMem, void *confetti
 	struct DrawConfettiScratch *scratch = CTR_SCRATCHPAD_PTR(struct DrawConfettiScratch, 0x30);
 	struct DrawConfettiTrigPair cameraTrig;
 	u32 screenBounds;
-	u_long *otBase;
+	uint32_t *otBase;
 	u32 currentParticles;
 	u32 particleCount;
 	s32 centerX;
@@ -332,14 +332,14 @@ void DrawConfetti(struct PushBuffer *pb, struct PrimMem *primMem, void *confetti
 				if (DrawConfetti_BoundsVisible(sxy0, sxy1, sxy2, sxy3, screenBounds))
 				{
 					u32 depth;
-					u_long *ot;
+					uint32_t *ot;
 
 					CtrGpu_WritePackedXY(&prim->x1, MFC2(12));
 					CtrGpu_WritePackedXY(&prim->x2, MFC2(13));
 					CtrGpu_WritePackedXY(&prim->x3, MFC2(14));
 					depth = MFC2(24);
 					CtrGpu_WriteColorCode(&prim->r0, scratch->color);
-					ot = (u_long *)(void *)((char *)otBase + ((depth >> 18) << 2));
+					ot = (uint32_t *)(void *)((char *)otBase + ((depth >> 18) << 2));
 					DrawConfetti_LinkPrimitive(prim, ot);
 					particleCount = scratch->remainingParticles;
 					prim++;
