+++ game/DrawSky/DrawSky.c
@@ -3,7 +3,7 @@
 struct DrawSkyContext
 {
 	const struct ShortVertex *verts;
-	u_long *ot;
+	uint32_t *ot;
 	u32 screenBounds;
 };
 
@@ -58,7 +58,7 @@ static void DrawSky_LoadFaceVertices(struct DrawSkyContext *ctx, const struct Sk
 	MTC2(DrawSky_ReadWord(&c->Color, 0x0), 22);
 }
 
-static void DrawSky_EmitPrimitive(u32 **primCursor, u_long *ot)
+static void DrawSky_EmitPrimitive(u32 **primCursor, uint32_t *ot)
 {
 	POLY_G3 *poly = (POLY_G3 *)*primCursor;
 
@@ -84,7 +84,7 @@ static u32 *DrawSky_Piece(struct Skybox *skybox, struct DrawSkyContext *ctx, int
 
 	for (u32 i = 0; i < numFaces; i++, face++)
 	{
-		u_long *ot = (u_long *)(void *)((char *)ctx->ot + (s16)face->D);
+		uint32_t *ot = (uint32_t *)(void *)((char *)ctx->ot + (s16)face->D);
 		u32 sxy0;
 		u32 sxy1;
 		u32 sxy2;
