+++ game/Particle.c
@@ -541,7 +541,7 @@ struct ParticleRenderListScratch
 		};
 	};
 	u8 pad_14[0x0c];
-	u_long *ot;
+	uint32_t *ot;
 	s32 cameraOffset[3];
 	s32 depth;
 };
@@ -614,16 +614,16 @@ static struct ParticleRenderListTrig Particle_RenderList_ReadTrig(s32 angle)
 	return trig;
 }
 
-static void Particle_RenderList_LinkPrimitive(u32 *tagWord, const void *packet, u_long *ot, u32 tag)
+static void Particle_RenderList_LinkPrimitive(u32 *tagWord, const void *packet, uint32_t *ot, u32 tag)
 {
 	CtrGpu_LinkPacket24(ot, tagWord, packet, tag);
 }
 
 static void Particle_RenderList_LinkAndAdvance(u32 **primCursor, u32 **payloadCursor, struct Particle *particle, struct InstDrawPerPlayer *idpp,
-                                               u16 flagsSetColor, s32 depth, u_long *defaultOT)
+                                               u16 flagsSetColor, s32 depth, uint32_t *defaultOT)
 {
 	u32 *prim = *primCursor;
-	u_long *otBase;
+	uint32_t *otBase;
 	s32 otIndex;
 
 	if (idpp != NULL)
@@ -636,7 +636,7 @@ static void Particle_RenderList_LinkAndAdvance(u32 **primCursor, u32 **payloadCu
 		if ((u16)idpp->depthOffset[1] < otIndex)
 			otIndex = (u16)idpp->depthOffset[1];
 
-		otBase = (u_long *)(uintptr_t)idpp->otRangeNormal;
+		otBase = (uint32_t *)(uintptr_t)idpp->otRangeNormal;
 	}
 	else
 	{
